+++ src/components/components.ts
@@ -1,4 +1,4 @@
-/* Web components. File generated at Thu, 30 Jul 2026 22:24:07 GMT. */
+/* Web components. File generated at Sun, 02 Aug 2026 02:59:46 GMT. */
 export * from './calc-prompt/calc-prompt.component';
 export * from './calc-prompt-list/calc-prompt-list.component';
 export * from './calculator-keypad/calculator-keypad.component';